Add code to handle a message sent over ethernet
After this change, tcp_echoserver_recv will use runCommsFSM to handle
messages sent over ethernet. Currently, eth_client.py sends a message
which should be type 0x00, but only gets the headers in response. No
idea why this is, but I need food and can't continue to work on this.

* Update olympus/drivers/comms.c/h
* Update olympus/main.c
* Update olympus/peripherals/ETH/tcp_echoserver.c/h
* Update olympus/eth_client.py