Fix headlights, turn signals, and misc lights
This change moves the GPIO ports around and enables the proper timers to get
the headlights, turn signals, and misc lights on the proper ports. It
also adds index inversion on MISC 1-4 (since TIM16 and TIM17 on MISC 5
and 6 already do this for some reason) to make the "set_headlights" and
"set_misc_lights" work on brightness and not "darkness" like it was
before (i.e. set_misc_lights(255) should not be off).

* Update Makefile
* Update lights.c/h